Install various addons - problematic.
This successfully adds the frontend and backend components of eea kitkat
addons package, as well as several other addons. However, there is a
peculiar problem: some standard blocks (eg, html block) are unavailable
with the kitkat package. Oddly, when the constituent addon blocks are
included by enumerating the individuals, rather than via the kitkat
collection, this problem does not occur.